Don't zap scons DB, handle release2 prebuilt case

+++ launch_chffrplus.sh
@@ -101,10 +101,11 @@ function launch {
       cp "$DIR/scripts/continue.sh" "/data/data/com.termux/files/continue.sh"
     fi
 
-    echo "Clearing build products and resetting scons state prior to NEOS update"
-    cd $BASEDIR && scons --clean
-    rm -f .sconsign.dblite
-    rm -rf /tmp/scons_cache
+    if [ ! -f "$DIR/prebuilt" ]; then
+      echo "Clearing build products and resetting scons state prior to NEOS update"
+      cd $BASEDIR && scons --clean
+      rm -rf /tmp/scons_cache
+    fi
     "$DIR/installer/updater/updater" "file://$DIR/installer/updater/update.json"
   fi
